Mon, Jan 17, 2022, 11:45 PM +00:00

+++ credit/credit.c
@@ -119,7 +119,7 @@ int main(void)
 	}
 
 
-	long sum_not_multiplied = first;
+	int sum_not_multiplied = first;
 	printf("Sum of non-multiplied: %i", sum_not_multiplied)
